adds dataset, analysis scripts, and updates docs
this adds the final dataset as CSV file, alongside
the python script used to analyze it and to create
visuals for the paper.

The README is extended to provide a starting point
and inform about the licensing.

As the paper is not yet available to the public,
no link and references are provided and will be
added in the future.

Signed-off-by: Tilmann Unte <unte@es-augsburg.de>